refactor(logs): move logs fixture to the appropriate directory
