+++ src/components/cc-logs-control/cc-logs-control.stories.js
@@ -2,7 +2,7 @@ import { makeStory } from '../../stories/lib/make-story.js';
 import '../cc-input-text/cc-input-text.js';
 import './cc-logs-control.js';
 
-import { createFakeLogs, CUSTOM_METADATA_RENDERERS } from '../cc-logs/fake-logs.js';
+import { createFakeLogs, CUSTOM_METADATA_RENDERERS } from '../../stories/fixtures/logs.js';
 
 export default {
   tags: ['autodocs'],
