+++ src/components/cc-logs/cc-logs.stories.js
@@ -5,7 +5,7 @@ import '../cc-input-number/cc-input-number.js';
 import '../cc-toggle/cc-toggle.js';
 import './cc-logs.js';
 
-import { CUSTOM_METADATA_RENDERERS, createFakeLogs } from './fake-logs.js';
+import { CUSTOM_METADATA_RENDERERS, createFakeLogs } from '../../stories/fixtures/logs.js';
 
 export default {
   tags: ['autodocs'],
@@ -15,7 +15,6 @@ export default {
 
 /**
  * @typedef {import('./cc-logs.js').CcLogs} CcLogs
- * @typedef {import('./cc-logs.types.js').Log} Log
  */
 
 const conf = {
